+++ packages/@tinacms/app/appFiles/src/main.tsx
@@ -13,11 +13,12 @@ limitations under the License.
 
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
+import styles from './index.css'
 import App from './App'
 
 ReactDOM.render(
   <React.StrictMode>
+    <style>{styles}</style>
     <App />
   </React.StrictMode>,
   document.getElementById('root')