+++ packages/@tinacms/schema-tools/src/types/SchemaTypes.ts
@@ -30,6 +30,9 @@ export type UIField<F extends UIField = any, Shape = any> = {
     meta: any,
     field: UIField<F, Shape>
   ): string | object | undefined | void
+  /**
+   * @deprecated use `defaultItem` at the collection level instead
+   */
   defaultValue?: Shape
 }
 
@@ -74,48 +77,66 @@ type Document = {
     extension: string
   }
 }
+export type TinaIndex = {
+  name: string
+  fields: {
+    name: string
+  }[]
+}
+
+export interface UICollection {
+  filename?: {
+    slugify?: (values: Record<string, any>) => string
+    readonly?: boolean
+  }
+  /**
+   * Forms for this collection will be editable from the global sidebar rather than the form panel
+   */
+  global?: boolean | { icon?: any; layout: 'fullscreen' | 'popup' }
+  /**
+   * Provide the path that your document is viewable on your site
+   *
+   * eg:
+   * ```ts
+   * router: ({ document }) => {
+   *   return `blog-posts/${document._sys.filename}`;
+   * }
+   * ```
+   */
+  router?: (args: {
+    document: Document
+    collection: TinaCloudCollection<true>
+  }) => string | undefined
+}
+
+type DefaultItem<ReturnType> = ReturnType | (() => ReturnType)
 
 interface BaseCollection {
   label?: string
   name: string
   path: string
+  defaultItem?: DefaultItem<Record<string, any>>
+  indexes?: TinaIndex[]
   format?: FormatType
-  ui?: {
-    /**
-     * Forms for this collection will be editable from the global sidebar rather than the form panel
-     */
-    global?: boolean | { icon?: any; layout: 'fullscreen' | 'popup' }
-    /**
-     * Provide the path that your document is viewable on your site
-     *
-     * eg:
-     * ```ts
-     * router: ({ document }) => {
-     *   return `blog-posts/${document._sys.filename}`;
-     * }
-     * ```
-     */
-    router?: (args: {
-      document: Document
-      collection: TinaCloudCollection<true>
-    }) => string | undefined
-  }
+  ui?: UICollection
   match?: string
 }
 
-type CollectionTemplates<WithNamespace extends boolean> =
+export type CollectionTemplates<WithNamespace extends boolean> =
   WithNamespace extends true
     ? CollectionTemplatesWithNamespace<WithNamespace>
     : CollectionTemplatesInner<WithNamespace>
 
+export type TinaTemplate = Template<false>
+
 interface CollectionTemplatesInner<WithNamespace extends boolean>
   extends BaseCollection {
-  templates: (string | Template<WithNamespace>)[]
+  templates: (string | GlobalTemplate<WithNamespace>)[]
   fields?: undefined
 }
 export interface CollectionTemplatesWithNamespace<WithNamespace extends boolean>
   extends BaseCollection {
-  templates: (string | Template<WithNamespace>)[]
+  templates: (string | GlobalTemplate<WithNamespace>)[]
   fields?: undefined
   references?: ReferenceType<WithNamespace>[]
   namespace: WithNamespace extends true ? string[] : undefined
@@ -155,11 +176,12 @@ export type TinaFieldEnriched = TinaFieldInner<true> & {
   parentTypename?: string
 }
 
-interface TinaField {
+export interface TinaField {
   name: string
   label?: string
   description?: string
   required?: boolean
+  indexed?: boolean
   // list?: boolean
   /**
    * Any items passed to the UI field will be passed to the underlying field.
@@ -208,7 +230,7 @@ type StringField =
       isBody?: boolean
       list: true
       isTitle?: never
-      ui?: UIField<any, string[]>
+      ui?: UIField<any, string[]> & { defaultItem?: DefaultItem<string> }
     }
 
 type BooleanField =
@@ -322,6 +344,7 @@ interface ObjectTemplatesInnerWithList<WithNamespace extends boolean>
           key?: string
           label?: string
         }
+        defaultItem?: DefaultItem<Record<string, any>>
       } & UIField<any, string>)
 }
 interface ObjectTemplatesInnerWithoutList<WithNamespace extends boolean>
@@ -353,6 +376,13 @@ interface ObjectTemplatesWithNamespace<WithNamespace extends boolean>
   type: 'object'
   visualSelector?: boolean
   required?: false
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
   list?: boolean
   /**
    * templates can either be an array of Tina templates or a reference to
@@ -375,6 +405,13 @@ interface InnerObjectFields<WithNamespace extends boolean> extends TinaField {
   type: 'object'
   visualSelector?: boolean
   required?: false
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
   /**
    * fields can either be an array of Tina fields, or a reference to the fields
    * of a global template definition.
@@ -391,6 +428,13 @@ interface InnerObjectFieldsWithNamespace<WithNamespace extends boolean>
   type: 'object'
   visualSelector?: boolean
   required?: false
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
   /**
    * fields can either be an array of Tina fields, or a reference to the fields
    * of a global template definition.
@@ -413,14 +457,14 @@ export type GlobalTemplate<WithNamespace extends boolean> =
     ? {
         label: string
         name: string
-        ui?: object | (UIField<any, any> & { previewSrc: string })
+        ui?: UICollection
         fields: TinaFieldInner<WithNamespace>[]
         namespace: WithNamespace extends true ? string[] : undefined
       }
     : {
         label: string
         name: string
-        ui?: object | (UIField<any, any> & { previewSrc: string })
+        ui?: UICollection
         fields: TinaFieldInner<WithNamespace>[]
       }
 
@@ -462,6 +506,13 @@ export type CollectionTemplateableObject = {
   namespace: string[]
   type: 'object'
   visualSelector?: boolean
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
   required?: false
   template: Templateable
 }