+++ experimental-examples/unit-test-example/CHANGELOG.md
@@ -4,6 +4,71 @@
 
 ### Patch Changes
 
+- tinacms@0.69.16
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [0c8c571d5]
+  - tinacms@0.69.15
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [c6e3bd321]
+  - tinacms@0.69.14
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [ea4a8e1b0]
+  - tinacms@0.69.13
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [183249b11]
+- Updated dependencies [8060d0949]
+  - tinacms@0.69.12
+
+## null
+
+### Patch Changes
+
+- tinacms@0.69.11
+
+## null
+
+### Patch Changes
+
+- tinacms@0.69.10
+
+## null
+
+### Patch Changes
+
+- tinacms@0.69.9
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [0513ae416]
+- Updated dependencies [64c40e6fc]
+- Updated dependencies [f3439ea35]
+- Updated dependencies [48032e2ba]
+- Updated dependencies [112b7271d]
+- Updated dependencies [8688dbff9]
+  - tinacms@0.69.8
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [9183157c4]
 - Updated dependencies [4adf12619]
   - tinacms@0.69.7